code cleanup done
